Store whether a column is valid or not
Instead of checking whether a column is valid every time it does need to
know just update the valid when a bit is applied (which is the only
place where it changes the column). And there it does need to know it
anyway. Might not be a big difference in JS, but in a Mikrocontroller it
probably saves a bit of computation to just look a value up every time a
column flashes.